型を変えた

+++ ble2.py
@@ -41,7 +41,7 @@ while True:
         if beacon.split(',')[1] == 'e7d61ea3f8dd49c88f2ff2484c07acb9' and beacon.split(',')[2] == '1' and beacon.split(',')[3] == '81':
             print "----------"
             rssi = beacon.split(',')[5]
-            print str(elapsed_time) + 's : ' + rssi + ',' + str(10**((-70 - int(rssi)) / 20))
+            print str(elapsed_time) + 's : ' + rssi + ',' + str(10**((-70 - float(rssi)) / 20))
             now = datetime.datetime.now()
             nowtime = now.strftime('%H:%M:%S.%f')
-            datalist.append(str(nowtime) + ',' + rssi + ',' + str(10**((-70 - int(rssi)) / 20)) + '\n')
+            datalist.append(str(nowtime) + ',' + rssi + ',' + str(10**((-70 - float(rssi)) / 20)) + '\n')
